refactor(metrics-collector): migrate seperate functions to internal
